post

+++ infras/src/main/java/com/pythongong/community/infras/web/AuthUserInfo.java
@@ -1,5 +1,5 @@
 package com.pythongong.community.infras.web;
 
-public record AuthUserInfo(Integer userId, String userType) {
+public record AuthUserInfo(Long userId, String userType) {
 
 }
